requirements.txt
# NLP
nltk==3.8.1

# Database
mysql-connector-python==8.3.0

# API layer (optional)
fastapi==0.110.0
uvicorn==0.29.0

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.6.4